Split NIRSpec IFU regtests
