+++ jwst/regtest/test_nirspec_ifu_spec3.py
@@ -12,42 +12,6 @@ TRUTH_PATH = 'truth/test_nirspec_ifu'
 
 
 @pytest.fixture(scope='module')
-def run_spec2(jail, rtdata_module):
-    """Run the Spec2Pipeline on a single exposure"""
-    rtdata = rtdata_module
-
-    # Setup the inputs
-    asn_name = 'single_nrs1_ifu_spec2_asn.json'
-    asn_path = INPUT_PATH + '/' + asn_name
-
-    # Run the pipeline
-    step_params = {
-        'input_path': asn_path,
-        'step': 'calwebb_spec2.cfg',
-        'args': [
-            '--steps.bkg_subtract.save_results=true',
-            '--steps.assign_wcs.save_results=true',
-            '--steps.imprint_subtract.save_results=true',
-            '--steps.msa_flagging.save_results=true',
-            '--steps.extract_2d.save_results=true',
-            '--steps.flat_field.save_results=true',
-            '--steps.srctype.save_results=true',
-            '--steps.straylight.save_results=true',
-            '--steps.fringe.save_results=true',
-            '--steps.pathloss.save_results=true',
-            '--steps.barshadow.save_results=true',
-            '--steps.photom.save_results=true',
-            '--steps.resample_spec.save_results=true',
-            '--steps.cube_build.save_results=true',
-            '--steps.extract_1d.save_results=true',
-        ]
-    }
-
-    rtdata = rt.run_step_from_dict(rtdata, **step_params)
-    return rtdata
-
-
-@pytest.fixture(scope='module')
 def run_spec3(jail, run_spec2):
     """Run the Spec3Pipeline on the results from the Spec2Pipeline run"""
     rtdata = run_spec2
@@ -102,17 +66,6 @@ def run_spec3_multi(jail, rtdata_module):
 
     rtdata = rt.run_step_from_dict(rtdata, **step_params)
     return rtdata
-
-
-@pytest.mark.bigdata
-@pytest.mark.parametrize(
-    'suffix',
-    ['assign_wcs', 'cal', 'flat_field', 'imprint_subtract', 'msa_flagging', 'pathloss', 'photom', 's3d', 'srctype', 'x1d']
-)
-def test_spec2(run_spec2, fitsdiff_default_kwargs, suffix):
-    """Regression test matching output files"""
-    rt.is_like_truth(run_spec2, fitsdiff_default_kwargs, suffix,
-                     truth_path=TRUTH_PATH)
 
 
 @pytest.mark.skip(reason="single exposure processing is redundant with calspec2")
